lib/rs/gf, lib/store/storedir: add missing type casts to support 32-bit machines

+++ lib/rs/gf/gf_test.go
@@ -15,12 +15,12 @@ func TestMInverseRaise(t *testing.T) {
 		inv := MInverse(v)
 		if (uint64(v)*uint64(inv))%MaxVal != 1 {
 			t.Errorf("MInverse(%v) = %v, but %v*%v %% %v = %v",
-				v, inv, v, inv, MaxVal, (v*inv)%MaxVal)
+				v, inv, v, inv, uint32(MaxVal), (v*inv)%MaxVal)
 		}
 
 		if inv != Raise(v, MaxVal-2) {
 			t.Errorf("Raise(%v, %v) = %v, wanted %v",
-				v, MaxVal-2, Raise(v, MaxVal-2), inv)
+				v, uint32(MaxVal-2), Raise(v, MaxVal-2), inv)
 		}
 	}
 }
